fix: add MainLayout file
