+++ package.json
@@ -43,27 +43,27 @@
     "@reduxjs/toolkit": "^2.5.1",
     "axios": "^1.8.2",
     "dayjs": "^1.11.13",
-    "expo": "~52.0.36",
+    "expo": "~52.0.42",
     "expo-blur": "~14.0.3",
-    "expo-constants": "~17.0.7",
-    "expo-dev-client": "~5.0.12",
+    "expo-constants": "~17.0.8",
+    "expo-dev-client": "~5.0.19",
     "expo-doctor": "^1.12.7",
     "expo-font": "~13.0.4",
     "expo-haptics": "~14.0.1",
     "expo-linking": "~7.0.5",
-    "expo-router": "~4.0.17",
+    "expo-router": "~4.0.20",
     "expo-splash-screen": "~0.29.22",
     "expo-status-bar": "~2.0.1",
     "expo-symbols": "~0.2.2",
-    "expo-system-ui": "~4.0.8",
-    "expo-updates": "~0.27.1",
+    "expo-system-ui": "~4.0.9",
+    "expo-updates": "~0.27.4",
     "expo-web-browser": "~14.0.2",
     "i18next": "^24.2.2",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-hook-form": "^7.54.2",
     "react-i18next": "^15.4.1",
-    "react-native": "0.76.7",
+    "react-native": "0.76.9",
     "react-native-device-info": "^14.0.4",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-modal-datetime-picker": "^18.0.0",
@@ -97,7 +97,7 @@
     "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-unused-imports": "^4.1.4",
     "jest": "^29.2.1",
-    "jest-expo": "~52.0.5",
+    "jest-expo": "~52.0.6",
     "jest-junit": "^16.0.0",
     "prettier": "^3.4.2",
     "react-test-renderer": "18.3.1",
